+++ PersistencePlugin/src/com/elmakers/mine/bukkit/plugins/persistence/dao/CommandData.java
@@ -1,4 +1,4 @@
-package com.elmakers.mine.bukkit.plugins.persistence.messages.dao;
+package com.elmakers.mine.bukkit.plugins.persistence.dao;
 
 import org.bukkit.command.Command;
 import org.bukkit.plugin.Plugin;
@@ -7,14 +7,32 @@ import org.bukkit.plugin.PluginDescriptionFile;
 import com.elmakers.mine.bukkit.plugins.persistence.annotations.PersistClass;
 import com.elmakers.mine.bukkit.plugins.persistence.annotations.Persist;
 
+/**
+ * A data class for encapsulating and storing a Command object.
+ * 
+ * @author nathan
+ *
+ */
 @PersistClass(name="command", schema="global")
 public class CommandData
 {
+	/**
+	 * The default constructor, used by Persistence to create instances.
+	 */
 	public CommandData()
 	{
 		
 	}
 	
+	/**
+	 * A constructor used to create new CommandData objects manually.
+	 * 
+	 * This may change in the future.
+	 * 
+	 * @param plugin The plugin that is registering this command
+	 * @param id The command id
+	 * @param command The command string, or alias
+	 */
 	public CommandData(Plugin plugin, String id, String command)
 	{
 		PluginDescriptionFile pdfFile = plugin.getDescription();
@@ -23,12 +41,29 @@ public class CommandData
 		this.command = command;
 	}
 	
+	/**
+	 * Check to see if this command matches a given command string.
+	 * 
+	 * Will eventually check permissions, look for sub-commands, and other 
+	 * things. 
+	 * 
+	 * @param cmd The command string to check
+	 * @return Whether or not the command succeeded
+	 */
 	public boolean checkCommand(Command cmd)
 	{
-		// TODO: permissions check, etc
+		// TODO: permissions check, sub-commands, etc
 		return command.equalsIgnoreCase(cmd.getName());
 	}
 	
+	/**
+	 * Get the command string that would be seen in a player message.
+	 * 
+	 * Currently, this is used only to generate help text- it will probably be
+	 * removed in the future.
+	 * 
+	 * @return The command name prefixed with a slash
+	 */
 	public String getCommandMatch()
 	{
 		return "/" + command;