+++ PersistencePlugin/src/com/elmakers/mine/bukkit/plugins/persistence/dao/PluginData.java
@@ -1,0 +1,14 @@
+package com.elmakers.mine.bukkit.plugins.persistence.dao;
+
+/**
+ * A class to encapsulate data for a plugin.
+ * 
+ * Each plugin can register any number of messages and commands.
+ * 
+ * @author nathan
+ *
+ */
+public class PluginData
+{
+
+}