+++ PersistencePlugin/src/com/elmakers/mine/bukkit/plugins/persistence/messages/dao/PluginData.java
@@ -1,6 +1,0 @@
-package com.elmakers.mine.bukkit.plugins.persistence.messages.dao;
-
-public class PluginData
-{
-
-}